pyproject.toml
[tool.poetry]
name = "python_scraping"
version = "0.1.0"
description = ""
authors = ["antiny <tranminhan.work@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.26.0"
selenium = "^3.141.0"
tqdm = "^4.62.1"
Pillow = "^8.3.1"
black = "^21.7b0"
python-string-utils = "^1.0.0"
typer = {extras = ["all"], version = "^0.3.2"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
